raspberrypi/requirements-pi.txt
# Core dependencies
websockets>=12.0
python-dotenv>=1.0.0
requests>=2.28.0

# Raspberry Pi GPIO support
RPi.GPIO>=0.7.0

# LCD I2C support
RPLCD>=1.3.0
smbus2>=0.4.0
